Add somewhat hacky wait for book request during transfer
Add somewhat hacky wait for book request while book transfer is in progress. The current return aborts processing of a valid book request that arrived nearly the same time as one from another Android device. They both *can* be processed as long as Desktop waits at this point for the transfer in progress to complete.