Refactor: put UDP and QR advertising in their own separate threads
I could not find a way to have the single file WiFiAdvertiser.cs make both advertisements show up actively at the same time. So I did a copy-paste-edit of WiFiAdvertiser, which is basically UDP, to create a QR version of it. Both advertisers are instantiated and started in WiFiPublisher.
The code is quite messy and the QR code presently doesn't get off its initial value, but it's late...