QR code now properly updates per current advert
When WiFiPublisher instantiates WiFiAdvertiserQR it now passes in a copy of its object reference so that WiFiAdvertiserQR can call the WiFiPublisher method to update the advert. WiFiPublisher has to act as the middle-man because the two advertisers can’t directly call into each other, since both are instantiated by WiFiPublisher as separate threads. But WiFiPublisher can see into both, so it pulls the current advert out of WiFiAdvertiser and then writes it into WiFiAdvertiserQR, who can then write the QR with it.
See 20240712_01_desktop_qr_update_refactor_ok.log + the two associated screenshots.