Suppress displaying a QR until UDP advert is valid
UDP and QR advertising are in separate threads so that both can be active at the same time. However the QR code needs to echo what is going out over UDP broadcast. The UDP advert may or may not be available when the QR thread is ready to display, so make the QR logic able to distinguish an invalid UDP advert and avoid displaying anything in that case.
Also do code cleanup.