All *Desktop* changes to date adding QR code scan and TCP to Bloom
My previous 'dev' branch got corrupted during an update. I was not able to fix it so I abandoned it and began work in this fresh branch. This commit is the first for this branch and represents 99% of what was in the corrupted branch. I say 99% because I did add a few more debug output stmts in efforts to learn why yesterday the Samsung tablet experienced 3 failures in a row to receive a book from a UDP advert. Today there is no such error...
Anyway, I am considering this a working UDP snapshot. Next I will remove the hacky text box I added months ago to simulate QR code content, then work to verify that QR coexists nicely with UDP and iron out any wrinkles.