Update WiFiAdvertiser.cs
First version of code that can get BD's SSID. Learning how to do this on a Windows computer has taken a surprising amount of time longer compared to Android.
This does work, for both wired and wireless connections, but it still has one main issue: it opens a dialog that the user should not have to see. Try to get the netsh process to not do that. Also, all the commented out attempts at this are still present for historical context. Will clean later.