Fix (hopefully) UDP receive problem
Two days ago I encountered, for the first time, a failure where Reader sent (repeatedly!) a valid book but Desktop did not receive it. Although I can't confirm, the most likely fault is the same as I did see on the sending/advertising side: the network stack used the wrong network interface.
This was possible on the receive side due to how UdpClient was constructed: only the port was specified, allowing the system to set the IP address. The current change here is to first create an endpoint with IP "0.0.0.0" (which tells the system to pay attention to ALL network interfaces) then pass that in to UdpClient's construction.
Preliminary testing looks good. Committing here before cleaning up debug output statements and other stuff.