Remove most debug output but prepare to mothball this branch
I recently learned that it *is* possible to force a UdpClient to use a specific IP address. That would be a smaller change to WiFiAdvertiser than the current replacement of UdpClient with Socket, and I want to try it. It will be in a different branch so need to get everything in this one committed for mothballing.