Desktop/Reader connectivity fix, UdpClient implementation
Works on Samsung tablet, wired and wifi. Need to test on BQ tablet then test at JAARS.
Trying this approach because I recently learned that, unlike what had thought, it *is* possible to instantiate a UdpClient with a specified IP address. Fixing the connectivity problem this way instead of replacing UdpClient with Socket will yield a pull request with fewer changes that should be easier to evaluate.